hp9k.c: Fixed regression (nw)
git-svn-id: svn://messdev.no-ip.org/mess@14921 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/drivers/hp9k.c
@@ -24,7 +24,7 @@ public:
 		m_p_videoram(*this, "p_videoram"){ }
 
 	required_device<cpu_device> m_maincpu;
-	required_shared_ptr<UINT8> m_p_videoram;
+	required_shared_ptr<UINT16> m_p_videoram;
 };
